refactor: introduce modular finite field parameterization (M31/M61)
- Added feature flags to `core_logic/Cargo.toml` to support the larger m61 field.
- Parameterized `FieldElement` and underlying field storage (`FieldStorage`) conditionally using M31 (31-bit Mersenne prime) or M61 (61-bit Mersenne prime).
- Implemented constant-time modular arithmetic, inversion, and conditional selection for both field representations.
- Adjusted sel4-compat pages, routing lorenz attractor seeding, and statistical tunnel simulation tests to dynamically adapt to the active modulus.

+++ core_logic/src/routing.rs
@@ -56,7 +56,7 @@ impl<const K: usize> HydraNode<K> {
         header_mac_key: FieldElement,
         header_nonce: FieldElement,
     ) -> Self {
-        let seed = node_index.value() ^ header_mac_key.value() ^ 0x13370000;
+        let seed = (node_index.value() ^ header_mac_key.value() ^ 0x13370000) as u32;
         HydraNode {
             node_index,
             trapdoor,
